Adicionado o ponto OpenCelliD.

+++ app/src/main/java/fernando/com/geolocationapicellid/Location2.java
@@ -3,8 +3,8 @@ package fernando.com.geolocationapicellid;
 public class Location2 {
     String status;
     int balance;
-    float lat;
-    float lon;
+    double lat;
+    double lon;
     int accuracy;
     String address;
     String message;
